Ajustes e refatoração

+++ src/test/java/com/mattutos/arkfuture/core/inventory/SimpleEnumContainerDataTest.java
@@ -18,7 +18,7 @@ class SimpleEnumContainerDataTest {
 
         assertEquals((short)0xffff, (short)containerData.get(EnumTest.SHORT));
         assertEquals(0xffff_ffff, (int)containerData.get(EnumTest.INT));
-        assertEquals(0xffff_ffff_ffff_ffffL, (long)containerData.get(EnumTest.LONG));
+        assertEquals(0xffff_ffff_ffff_ffffL, containerData.get(EnumTest.LONG));
     }
 
     @Test
